% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BivariateReport}
\alias{BivariateReport}
\title{Bivariate report}
\usage{
BivariateReport(data, dictionary, variable)
}
\arguments{
\item{data}{A data frame}

\item{dictionary}{A data dictionary}

\item{variable}{A stratifying variable}
}
\description{
This function takes in a formatted data frame, data dictionary, and stratifying 
variable to perform comparison test and analysis. Knows when to use tests
for categorical and quantitative variables.
}
