% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CategoricalTable}
\alias{CategoricalTable}
\title{Categorical variable summary}
\usage{
CategoricalTable(variable)
}
\arguments{
\item{variable}{Categorical variable}
}
\description{
This function takes in variable, returns counts, frequency, and missing information.
Used internally in report summary functions.
}
