% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Fixed}
\alias{Fixed}
\title{Format a decimal output}
\usage{
Fixed(num, digits = 1)
}
\arguments{
\item{num}{Numeric input}

\item{digits}{Number of digits to round to}
}
\description{
Function to round and format numerical output. Used implicitly in summary functions.
}
