% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatData.R, R/functions.R
\name{FormatData}
\alias{FormatData}
\title{Format a data frame according to data dictionary}
\usage{
FormatData(data, dictionary)

}
\arguments{
\item{data}{A data frame}

\item{dictionary}{A data dictionary}

}
\description{

This function takes in a data frame and formatted data dictionary. Formats
all variable entries (contiuous, categorical, etc) and returns formatted data frame.
}
