% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeTable.R, R/functions.R
\name{QuantitativeTable}
\alias{QuantitativeTable}
\title{Quantitative Table}
\usage{
QuantitativeTable(variable)

QuantitativeTable(variable)
}
\arguments{
\item{variable}{Quantitative variable}

\item{x}{Quantitative variable}
}
\description{
Takes in a quantitative variable and returns summary statistics and information.

Takes in a quantitative variable and returns summary statistics and information. 
Used internally in report summary functions.
}
