% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TextTable}
\alias{TextTable}
\title{Text Table}
\usage{
TextTable(variable)
}
\arguments{
\item{variable}{Variable}
}
\description{
Takes in character variable strings and returns percent complete.
}
