% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TrimP}
\alias{TrimP}
\title{Cleaned up output for p-values}
\usage{
TrimP(p)
}
\arguments{
\item{p}{p-value (numeric)}
}
\description{
Formatting p-values and categorizing as <0.001, 1, or its fixed value.
Used implicitly in summary functions.
}
