% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{UnivariateReport}
\alias{UnivariateReport}
\title{Univariate Report on data frame}
\usage{
UnivariateReport(data, dictionary)
}
\arguments{
\item{data}{Data frame}

\item{dictionary}{Data dictionary}
}
\description{
Takes in formatted data frame and data dictionary, returns a summary report of data.
Knows when to identify categorical and quantitative variables.
}
