% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VectorTable}
\alias{VectorTable}
\title{Vector Table}
\usage{
VectorTable(variables)
}
\arguments{
\item{variables}{Matrix of entires of vector variable}
}
\description{
Function for variables of the format "check all that apply." Returns
counts and frequencies per option.
}
